msm: 7x30: Remove maxx and maxy fields from ttsp platform data + cleanup
To suit the needs of 8660 fluid, maxx and maxy variables of platform
data are changed to panel_maxx and panel_maxy. This patch removes
the maxx and maxy variables and another patch will be submitted to
use the correct variables.

Also removed maxz which wasn't present either in Code Aurora board files.

Change-Id: I789a9bf8cedd9289d09e5d2083db4bb005afc85c
Signed-off-by: Mohan Pallaka <mpallaka@codeaurora.org>
Signed-off-by: Adrien Tétar <adrien.tetar@hotmail.fr>