msm: kgsl: Move pagetable allocation to a DMA pool
Instead of using a single coherent DMA allocation of memory at
boot time, create and use a DMA pool to get memory for pagetables.
Since coherent DMA allocations max out at 2Mb, a reasonable number
of reasonably sized pagetables would max out the memory and the
KGSL driver would fail to load.  The downside is that using a pool
has a risk that we will be out of coherent memory when we need it,
so we make sure to raise a stink when it happens so it can be easily
diagnosed.

Change-Id: I40da9c65d4f924cb280d0b13ca823101d1cec6df
Signed-off-by: Jordan Crouse <jcrouse@codeaurora.org>