epoll: initialize slack for negative timeout values
When a negative timeout value is passed to epoll the 'slack' variable is
currently unitialized:

fs/eventpoll.c: In function â€˜ep_pollâ€™:
fs/eventpoll.c:1119: warning: â€˜slackâ€™ may be used uninitialized in this function

In this case a NULL pointer is passed to schedule_hrtimeout_range()
specifying an infinite timeout. The current implementation of
schedule_hrtimeout_range() does not use slack in this case, but we
should still initialize slack to 0 in case future implementations use it.

Signed-off-by: Shawn Bohrer <shawn.bohrer [at] gmail>
Acked-by: Davide Libenzi <davidel [at] xmailserver>